Major BOLD Preprocessing Update
Overhauled the preprocessing procedure so that it is invoked via a
static class method. Completely reorganized the preprocessing parameter
storage file so that it's easier to understand. Reworked almost all of
the preprocessing methods to be compatible with the parameter
adjustments. Tested all functions up through conditioning, so now all
that's left is testing that and writing/testing the slice-timing
correction procedure.

Also created a structure utilities package and wrote a new function to
merge two or more structures. Moved some other structure-related
functions there too.