Finished With Human Data Object Updates
Haven't needed to make any major changes to the human data objects in a
little while, so this will serve as the final commit before the merge
back into the main branch.

BOLD data preprocessing still isn't outputting results that agree with
previously preprocessed data. I don't know why this is, exactly, but I
suspect it has something to do with the new slice timing and/or motion
corrections that I implemented in SPM. I'll have to investigate this
later if time permits.

Specific Updates
- Wrote and documented a general function for converting Pearson
correlation coefficients to z-scores called fishersTransform.
- Improved the size checking between the data array and mask in the
function maskImageSeries.
- Implemented explicit autocorrelations in xcorrArr and moved an error
check so that it occurs earlier in the function.