Path & File Management Updates
Made several updates to the new Path and File classes. Also implemented
a new Today class to manage daily data and activity log scripts.

File Class Updates:
- Filled out the documentation for this class and all existing methods.
- Implemented basic data stream functionality for IO operations on
files.
- Implemented a destructor to cover the release of stream resources in
case of a crash or object deletion.
- Fixed a bug with using Path objects in the constructor method that
resulted in crashes.

Path Class Updates:
- Implemented a new method to convert Path objects to cell arrays.
- Implemented a new method to open directories in Windows Explorer.

Paths Class Update:
- Completely rewrote this so that it's a makeshift static class (true
static classes aren't allowed in MATLAB). This is a somewhat hacky
workaround, but static class properties can be achieved using static
methods, which makes using this class a lot easier. It also allows for
chaining operations with the new "properties", which couldn't be done
before unless an instance of the class was created first.
- Removed the constructor method since it's no longer needed. Other
methods will be removed in the near future as they are taken over by the
Path/File classes.

New Today Class:
- Wrote a new class to manage daily data and to provide utilities for
Today Scripts, which are essentially my daily activity logs in MATLAB
for work.
- So far, the static properties and script creation methods are written
and working.

Other Updates:
- Wrote the WarnAndReturn function to handle displaying warnings in the
console and forcing an early return from the calling function. This was
intended for use in functions where an error check fails, but not
necessarily catastrophically for the executing program. However, it
doesn't work at all because "return" is a keyword and can't be invoked
in different workspaces. I don't see any other way to produce this
effect without extra code, so it's likely this function will get deleted
in the future.
- Deleted xcorrVolume. It was never written and didn't need to be; I
found another way to accomplish that operation without writing a whole
new function.