Minor Updates & Reorganization
Path Class Updates:
- Implemented a new method called Contents to get both the file and
folder contents of a directory path.
- Changed the name of the method ViewInExplorer to just View so that
it's easier to type in the command line.
- Removed the class folder that contained this object since it wasn't
needed. This class now resides directly in the File Management folder.

File Class Updates:
- Removed the class folder that contained this object since it wasn't
needed. This class now resides directly in the File Management folder.

Paths Class Update:
- Removed the static path to DataObjects and replaced it with a path to
Data, which points to the general data folder on my flash drive.

Other Updates:
- Updated the call to Paths.DataObjects in the "startup" script to
maintain compatibility.
- Made the function "empiricalcdf" use GPU processing because benchmarks
showed that the original block-vectorized processing was a terrible idea
performance-wise. This will have to be updated in the future though to
be useful on computers without usable GPUs.