New Input Assignment Function
Created a new function called "assignto" that is meant to replace
"assignInputs". This function improves on the previous method by not
requiring a structure that hides variable initialization from the code
analyzer (which has been a constant source of frustration for a long
time now). It additionally strips away the variable name compatibility
that "assignInputs" had that was just bad programming practice and went
unused most of the time anyway. Lastly, it also does away with automatic
variable formatting, which again was unused most of the time and should
probably be a part of the function logic to begin with.

This new system has not yet been implemented in any functions and has
only undergone minimal testing for its functionality. It seems to work
pretty well though and is much more elegant than the previous variable
assignment method.