New Montage Class
Wrote a new class called Montage, which is basically a more general
version of BrainPlot that creates montages of MATLAB axes objects. It
can be used to make montages just like BrainPlot, but is much more
flexible. Documentation still needs to be improved, however.

To begin, I wrote some class methods to handle basic 2D plots. This
includes a general Plot method as well as a ShadedPlot method, which
additionally draws some shading around a line. I also wrote in a method
that allows any individual montage element to be "expanded" when clicked
on, which basically involves showing a larger version of same plot in
its own new window.