Today Class Updates
Update details:
- Implemented a new method that performs a basic archiving of Today Data
that's stored on my portable flash drive. It copies any folders off of
the flash drive into the data archive on the computer I'm using,
provided a folder by that name (the date name of the script that
generated it) doesn't already exist there. This really needs a method
for recursively scanning files and folders to transfer any new or
different data over, but this will do as a start.
- Implemented a method for automatically loading global variables listed
at the beginning of a Today Script.
- Implemented some static fields for storing date and time formatting
strings in case I want to change these in the future.
- Implemented some standard assertions for date- and time-related
strings.
- Removed automatic placement of infraslow data set references in
section text. These weren't being used enough to justify having them
there.