Montage Class Updates
Update Details:
- Changed the y-axis tick labels to be automatically flipped by the
getter/setter methods so that labeling starts at the upper left and
continues downward towards the axes origin. This was done because this
is the same ordering in which the montage is filled in and because I
keep making the mistake of not flipping the labels manually.
- Fixed some bugs related to automatic titling in the expanded element
figures.