New Function "wassert"
Wrote a new function called "wassert" that works just like assert except
that it only issues a warning in the console when its condition fails
instead of throwing an error that crashes programs. This is useful for
informing users of potential problems even though the program can
technically handle whatever condition it found.