Montage Class Updates
Implemented a new static constructor method called "EEG" that creates
montages of electrode spatial mappings and colors them in.

This is essentially a much more efficient version of calling the
BrainPlot class with EEG data, which I'm planning to phase out. Using
the Montage class also allows for the reuse of montages between similar
successive data sets. Reusing montages is a lot faster because the setup
overhead is eliminated. This is especially influential here because the
channel mappings must be loaded off of the hard drive and copied across
elements of the montage.