New Warning Function & Deprecation Tag
Created a new function that allows warnings to be issued as if they
originated from a calling function. This is not natively available with
MATLAB warning functions. Also implemented a new function called
DEPRECATED, which serves as a flag that's meant to be placed in old
functions that will be removed soon. It throws warnings to remind users
not to continue using the functions in which its placed.