Empirical CDF Function Updates
Changed some of the C code so that it works on the assumption that the
null data distribution has been sorted before being inputted. This
should increase the speed of CDF generation a little bit. The MATLAB
code wrapper now sorts null values before handing them off to the MEX
function.