Thresholding Function Rewrite
Completely rewrote the function "threshold" to improve its performance
and get rid of a ton of junk code that no longer worked and/or had been
commented out ages ago. This function now exclusively operates on
empirically generated data distributions (real and null) and no longer
requires that p-values be generated for all of the data. It also now
uses the SGoF G-test for FWER correction by default instead of the
binomial method.