Added Deprecated Code to Repository
Ordinarily I store my deprecated code outside of my usual MATLAB
directory to keep it all out of the way. However, with the creation of
the lab's Git repository I figured I needed add it in. Some of my old
work makes use of a number of these functions, and if anyone had tried
to follow it using only the in-use, versioned code from my repository,
they would hit some dead ends that might be inconvenient.

Unfortunately, some of the deprecated functions I used to use got
deleted or updated so much that they would be non-functional anyway.
Thus, this won't solve all of the problems, but it's the best I can do
for now.