Committing Code Prototyping to Repository
This is the initial commit of my existing code prototyping folder. The
current contents of this folder are mostly related to prototyping some
new functionality in MATLAB and in C and for benchmarking attempts at
improving some computationally intense routines that I regularly have to
run.

This folder also contains a number of MEX binaries that are prototype
builds of new C functions. Like my other MEX functions, these are built
with the Intel optimizing compiler and should not be run unless the
necessary dependencies are in place on your system.