File Management Package Updates
Completely overhauled the documentation across all code in this folder.
Class documentation now summarizes all of the properties and methods
that are available in the console window.

Also rewrote most of the Path class so that it now serves as an abstract
base for File and the newly implemented Folder objects. The Folder class
assumes much of the former functionality that Path previously had.