Removed Color Restrictions from Window Class
I'm removing certain restrictions that previously limited Background and
Colormap values to be of type Color for the Window class. Although nice
in principle, there are too many glaring problems with this approach.

One problem is that the use of Color data structures isn't going to be
intuitive to anyone but me (MATLAB doesn't have a Color type). A bigger
problem, however, is that I haven't been able to address the serious
performance issues with the Color data class, particularly when used
like bitmaps. Things can slow down to an unbearable degree, particularly
in the HG2 system, and I'm not sure I'll be able to solve the problem
any time soon.

Thus, for now at least, the Window class will go back to using ordinary
numeric arrays for these properties.