Added some documentation and changed window closing behavior.
The 'Close' method no longer attempts to destroy the window handle
variable in the calling workspace. This would have been nice to have
in place so that accidentally utilizing a deleted window handle
wouldn't be possible, but there's no way to do this across multiple
calling scopes (i.e. if a window handle is passed through several
functions before being closed), so I've just decided to leave that
kind of resource management up to the user.